requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
flatbuffers==25.12.19
gunicorn
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
llvmlite==0.46.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
onnxruntime==1.24.4
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
pooch==1.9.0
protobuf==7.34.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyMatting==1.1.15
pytest==9.0.2
python-dotenv==1.2.2
python-multipart==0.0.22
referencing==0.37.0
rembg==2.0.73
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
starlette==0.52.1
sympy==1.14.0
tifffile==2026.3.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
